bot/serializer.py
# coding: utf-8

from rest_framework import serializers

from .models import Message


class MessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = Message
        fields = ('mid', 'content')
