requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
cffi==1.7.0
cryptography==1.4
decorator==4.0.10
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.7
django-filter==0.13.0
django-toolbelt==0.0.1
djangorestframework==3.3.3
gnureadline==6.3.3
google-api-python-client==1.5.1
gspread==0.3.0
gunicorn==19.6.0
httplib2==0.9.2
idna==2.1
ipython==4.2.0
ipython-genutils==0.1.0
oauth2client==1.5.2
pexpect==4.1.0
pickleshare==0.7.2
psycopg2==2.6.1
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==16.0.0
requests==2.10.0
rsa==3.4.2
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
static3==0.7.0
traitlets==4.2.1
uritemplate==0.6
whitenoise==3.2
